refactor(core): move methods out from file_ctl - move  parse_json to json module - move parse_http_packet to http module
