+++ fhe_http_core/src/utils/http.rs
@@ -1,15 +1,19 @@
 use crate::fhe_traits::key_serialize::KeySerialize;
 use crate::utils::base64;
-use crate::utils::file_ctl;
 use crate::utils::json;
+use serde_json::Value;
+use tfhe;
 use tfhe::ClientKey;
 
 pub fn create_fhe_header(method: &str) -> String {
-    return file_ctl::create_fhe_header(&method);
+    let mut header = serde_json::Map::new();
+    header.insert("fhe-method".to_string(), Value::String(method.to_string()));
+
+    return serde_json::to_string(&header).unwrap();
 }
 
 pub fn encrypt_fhe_body(keys: Vec<String>, data: &str, client_key: Vec<u8>) -> String {
-    let body = file_ctl::parse_json(data);
+    let body = json::parse_json(data);
     let client_key: ClientKey = KeySerialize::deserialize(&client_key);
     let keys = keys.iter().map(|x| x.as_str()).collect();
     let encrypted_body = json::encrypt_json(&keys, &body, &client_key);
@@ -25,27 +29,35 @@ pub fn decrypt_fhe_body(keys: Vec<String>, data: &str, client_key: Vec<u8>) -> S
 }
 
 pub fn set_server_key_in_body(server_key: &Vec<u8>, data: &str) -> String {
-    let mut body = file_ctl::parse_json(data);
+    let mut body = json::parse_json(data);
     body.insert(
         "server_key".to_string(),
         serde_json::Value::String(base64::encode_vec_u8(server_key)),
     );
     return serde_json::to_string(&body).unwrap();
 }
+pub fn parse_http_packet(packet: &str) -> (String, String) {
+    // split the packet into header and body
+    let res: Vec<&str> = packet.split("\r\n\r\n").collect();
+    if res.len() == 2 {
+        return (res[0].to_string(), res[1].to_string());
+    }
+    return ("".to_string(), "".to_string());
+}
 
 pub fn check_http_packet(packet: &str) -> Result<(), &str> {
-    let (header, body) = file_ctl::parse_http_packet(packet);
+    let (header, body) = parse_http_packet(packet);
     if header == "" || body == "" {
         return Err("Error parsing HTTP packet");
     }
     // check if fhe-method is present in the header
-    let header = file_ctl::parse_json(&header);
+    let header = json::parse_json(&header);
     if !header.contains_key("fhe-method") {
         return Err("fhe-method not found in the header");
     }
 
     // check if server_key is present in the body
-    let body = file_ctl::parse_json(&body);
+    let body = json::parse_json(&body);
     if !body.contains_key("server_key") {
         return Err("server_key not found in the body");
     }
